requirements.txt
# Python dependencies for Spotify verification script
# Install with: pip install -r requirements.txt

requests>=2.31.0
python-dotenv>=1.0.0  # Optional: for loading .env files automatically
